requirements.txt
alabaster==0.7.16
amqp==5.0.9
asgiref==3.8.1
auditlog3==1.0.1
babel==2.16.0
bcrypt==4.0.1
beautifulsoup4==4.13.3
billiard==3.6.4.0
bleach==6.0.0
celery==5.2.3
certifi==2021.10.8
cffi==1.17.1
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.5.4
cron-descriptor==1.2.35
cryptography==44.0.2
Deprecated==1.2.13
Django==4.2.20
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-email==3.0.0
django-jet-reboot==1.3.10
django-markdownify==0.9.3
django-markdownx==4.0.2
django-nested-admin==4.0.2
django-oauth-toolkit==3.0.1
django-split-settings==1.2.0
django-timezone-field==5.0
docutils==0.21.2
et_xmlfile==2.0.0
idna==3.3
imagesize==1.4.1
importlib-metadata==6.2.0
josepy==1.13.0
mozilla-django-oidc==3.0.0
importlib_resources==6.5.2
Jinja2==3.1.4
jwcrypto==1.5.6
kombu==5.2.3
Markdown==3.4.3
MarkupSafe==3.0.2
mysqlclient==2.1.0
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
Pillow==9.0.0
prompt-toolkit==3.0.24
pycountry==24.6.1
pycparser==2.22
Pygments==2.18.0
pymemcache==4.0.0
pyOpenSSL==23.2.0
pyparsing==3.0.6
pypdf==5.0.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-monkey-business==1.0.0
pytz==2021.3
redis==4.1.0
requests==2.32.3
rstr==3.2.2
rules==3.3
schwifty==2024.11.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==7.4.7
sphinxawesome-theme==5.3.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.4.2
tinycss2==1.1.1
tomli==2.1.0
typing_extensions==4.12.2
tzdata==2022.7
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
XlsxWriter==3.0.2
zipp==3.15.0
